requirements.txt
python-telegram-bot==21.6
motor==3.6.0
python-dotenv==1.0.1
dnspython==2.7.0
